rename datapipelines and worksheets repository files
