+++ backend/dataall/modules/datapipelines/cdk/datapipelines_cdk_cli_wrapper_extension.py
@@ -4,7 +4,7 @@ from dataall.base.aws.sts import SessionHelper
 from dataall.base.cdkproxy.cdk_cli_wrapper import CDKCliWrapperExtension, \
     describe_stack, update_stack_output
 from dataall.modules.datapipelines.cdk.datapipelines_cdk_pipeline import CDKPipelineStack
-from dataall.modules.datapipelines.db.repositories import DatapipelinesRepository
+from dataall.modules.datapipelines.db.datapipelines_repository import DatapipelinesRepository
 
 
 logger = logging.getLogger('cdksass')
